Basic mechanism to use a digital twin motor

+++ mcu-firmware/rrrc/components/MotorPortHandler/MotorPortHandler.h
@@ -42,7 +42,7 @@ typedef struct {
     float speed_limit;
     DriveRequest_RequestType_t request_type;
     DriveRequest_RequestValue_t request;
-    float positionBreakpoint;
+    uint32_t positionBreakpoint;
 } DriveRequest_t;
 typedef float Current_t;
 typedef float Percentage_t;
